requirements.txt
djangorestframework
certifi==2019.9.11
Django==2.2.5
gunicorn==20.0.0
nltk==3.4.5
numpy==1.17.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
